src/main/java/com/example/controller/GestionControllerTest.java
package com.example.controller;

import com.example.model.Gestion;
import com.example.service.GestionService;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.util.Collections;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@WebMvcTest(GestionController.class)
public class GestionControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private GestionService gestionService;

    @Test
    public void getAllGestionesTest() throws Exception {
        Mockito.when(gestionService.getAllGestiones()).thenReturn(Collections.emptyList());

        mockMvc.perform(get("/api/gestiones"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$").isArray())
                .andExpect(jsonPath("$.length()").value(0));
    }
}
